Second PR for EVSE code driven conversion
