+++ src/app/clusters/energy-evse-server/app_config_dependent_sources.gni
@@ -13,6 +13,8 @@
 # limitations under the License.
 app_config_dependent_sources = [
   "EnergyEvseCluster.cpp",
+  "EnergyEvseCluster.h",
+  "EnergyEvseDelegate.h",
   "EnergyEvseTestEventTriggerHandler.h",
   "energy-evse-server.h",
 ]
